requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
anthropic>=0.40.0
moviepy>=1.0.3
opencv-python>=4.8.0
# pystray>=0.19.0  # Requires Python 3.9+ - disabled for Python 3.8 compatibility
